Use remote coroutine worker
