+++ pir/pir-internal/src/main/java/com/duckduckgo/pir/internal/service/PirRemoteWorkerService.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2025 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.pir.internal.service
+
+import android.os.Process
+import androidx.work.multiprocess.RemoteWorkerService
+import logcat.AndroidLogcatLogger
+import logcat.LogPriority
+import logcat.LogcatLogger
+import logcat.logcat
+
+class PirRemoteWorkerService : RemoteWorkerService() {
+    override fun onCreate() {
+        super.onCreate()
+        LogcatLogger.install(AndroidLogcatLogger(LogPriority.DEBUG))
+        logcat { "PIR-WORKER: Running in process: ${Process.myPid()}" }
+    }
+}
