+++ pir/pir-internal/src/main/java/com/duckduckgo/pir/internal/service/PirSupportedBrokers.kt
@@ -1,0 +1,70 @@
+/*
+ * Copyright (c) 2025 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.pir.internal.service
+
+internal val supportedBrokers = listOf(
+    "AdvancedBackgroundChecks",
+    "backgroundcheck.run",
+    "Centeda",
+    "Clubset",
+    "ClustrMaps",
+    "Councilon",
+    "CurAdvisor",
+    "Cyber Background Checks",
+    "Dataveria",
+    "FastBackgroundCheck.com",
+    "FastPeopleSearch",
+    "FreePeopleDirectory",
+    "Inforver",
+    "Kwold",
+    "mylife",
+    "Neighbor Report",
+    "New England Facts",
+    "OfficialUSA",
+    "People Background Check",
+    "People-Wizard.com",
+    "PeopleFinders",
+    "People Search Now",
+    "PeoplesWhizr",
+    "PeoplesWiz",
+    "PeoplesWizard",
+    "PeopleWhiz.com",
+    "PeopleWhiz.net",
+    "PeopleWhized.com",
+    "PeopleWhized.net",
+    "PeopleWhizr.com",
+    "PeopleWhizr.net",
+    "PeopleWiz",
+    "PeopleWizard.net",
+    "PeopleWizr",
+    "Pub360",
+    "PublicReports",
+    "Quick People Trace",
+    "Search People FREE",
+    "SmartBackgroundChecks",
+    "Spokeo",
+    "TruePeopleSearch",
+    "Verecor",
+    "USA People Search",
+    "USA Trace",
+    "USPhoneBook",
+    "Vericora",
+    "Veriforia",
+    "Veripages",
+    "Virtory",
+    "Wellnut",
+)
